requirements.txt
babel
click
google-generativeai